add initial support for `changelog` & `changeloglock`
this also comes with various supporting infrastructure, e.g. the
executor.

this does not yet allow executing any changesets, but it allows
liquibase to run with an empty changeset which will create the
`changelog` and also manage the lock (which ensures that only one
liquibase operation is running at a time).

note: the `liquibase.nosql` package has been copied from
[`liquibase-mongodb`] and adapted where needed (it is not 100% generic).
no authorship is claimed for this content!

it's important that all document operations being performed on the
changelog & lock indices are done with `refresh=wait_for` to ensure that
a subsequent read will see the correct data (otherwise a changeset might
be marked as run but a subsequent run will not see this and try to run
it again).

[`liquibase-mongodb`]: https://github.com/liquibase/liquibase-mongodb