use testcontainers' `@Container` annotation
so far we started the testcontainer manually. however, [the official
way][quickstart] is to use the `@Testcontainers` and `@Container`
annotations.

the major difference in behaviour between the two is that in our
existing setup we started one container for all tests together, so they
could influence each other, while the new setup launches a new container
for each test. the downside of this is that test execution now takes
longer as it always needs to spin up a new container and wait for
OpenSearch to start. as we don't have that many tests this should be
acceptable and the improved stability is probably worth it.

[quickstart]: https://java.testcontainers.org/quickstart/junit_5_quickstart/