support multiple URLs in connection string
the underlying OpenSearch library supports multiple URLs, thus we should
do the same here as well.

since Liquibase has no concept of this on its own we have to use a
workaround and use a character to concatenate multiple URLs together.
`,` is being used for this as it's normally not part of URLs and it's
already used in the OpenSearch client for this purpose.
the `opensearch:` prefix should not be repeated.

i.e. a URL with multiple addresses should be defined like this:
```
opensearch:http://localhost:9200,http://localhost:9201,http://localhost:9203
```
the login information is the same for all, since they must form a
cluster.