allow creating `OpenSearchConnection` with existing client
so far we only supported creating the connection through the standard
liquibase mechanism, where liquibase would find it in its registry and
instantiate it. this is inadequate for a lot of use-cases since it's
often necessary to customise the client further than what's possible
with just a few parameters.

the current client creation disables TLS (since OpenSearch is usually
run with custom TLS certificates and we have no way of passing them in)
and only supports basic authentication.

by giving the option to pass in an instantiated client it's now possible
to use all forms of authentication, especially those which require a
custom callback to create an authentication token (e.g. JWT).

this will also be used by `liquibase-opensearch-spring-boot-starter` to
pass in an existing `OpenSearchClient`, usually created by
`spring-data-opensearch-starter`.

resolves #29 (indirectly, as it's now possible to pass in a client which
has its own custom authentication set up for JWT).