validate TLS certificates in `OpenSearchClient` created by us
this removes the `TrustAllStrategy` used so far to accept custom TLS
certificates.

this can now be done since we now have a way to create an
`OpenSearchConnection` with a custom `OpenSearchClient`, allowing the
caller to set it up in whichever way they need - including disabling TLS
validation.

this makes our default more strict, which is also more secure.