prepare for first release: add missing features & documentation and clean up code (#133)
* `OpenSearchConnection#getURL`: return cluster name if connected

it makes more sense to return the cluster name if we're connected. this
is especially useful when considering the next commit building on top of
this: there `uris` will no longer be set in some circumstances.

* allow creating `OpenSearchConnection` with existing client

so far we only supported creating the connection through the standard
liquibase mechanism, where liquibase would find it in its registry and
instantiate it. this is inadequate for a lot of use-cases since it's
often necessary to customise the client further than what's possible
with just a few parameters.

the current client creation disables TLS (since OpenSearch is usually
run with custom TLS certificates and we have no way of passing them in)
and only supports basic authentication.

by giving the option to pass in an instantiated client it's now possible
to use all forms of authentication, especially those which require a
custom callback to create an authentication token (e.g. JWT).

this will also be used by `liquibase-opensearch-spring-boot-starter` to
pass in an existing `OpenSearchClient`, usually created by
`spring-data-opensearch-starter`.

resolves #29 (indirectly, as it's now possible to pass in a client which
has its own custom authentication set up for JWT).

* validate TLS certificates in `OpenSearchClient` created by us

this removes the `TrustAllStrategy` used so far to accept custom TLS
certificates.

this can now be done since we now have a way to create an
`OpenSearchConnection` with a custom `OpenSearchClient`, allowing the
caller to set it up in whichever way they need - including disabling TLS
validation.

this makes our default more strict, which is also more secure.

* remove TODO about supporting other credential providers

this is no longer relevant since callers can now pass their own
`OpenSearchClient` and use whichever credentials provider they'd like.

* add content to README

* add test with security

so far we didn't test with security enabled. we keep it disabled since
now we validate the TLS certificate in the standard path (which we
mostly test), so we use the test with the custom client to also test
security (the client instantiation is otherwise the same).

* remove unnecessary TODOs

there's nothing left to do here. it's just a fact that these methods
exist but don't do anything.
this can be cleaned up only once liquibase/liquibase#4236 gets
implemented.

* implement `tag` method

more information can be found in the documentation: https://docs.liquibase.com/commands/utility/tag.html

* fix most sonarqube warnings

* update tests to OpenSearch 2.19.2

* simplify OpenSearch `Query` usage in tests

* add test coverage for OpenSearch 3

OpenSearch 3 was released a short while ago. since `opensearch-java` is
compatible with both OpenSearch 2.x and 3.x (at least for the
functionality we plan on using) we should have test coverage for both.

* README: add note about semantic versioning