update to `opensearch-testcontainers` 3.0.1
this contains two breaking changes in that library:
- they renamed `OpensearchContainer` to `OpenSearchContainer`
- they increased the minimum supported java version to 21

for us this is not a breaking change as it is purely a test dependency
and has no impact on our public API. we continue to support java 17,
however we can now only run our tests on java 21.

also the nightly build now also runs integration tests.